+++ S3ServerInterface/Callbacks/BucketCallbacks.cs
@@ -15,111 +15,111 @@ namespace S3ServerInterface
 
         /// <summary>
         /// Delete a bucket.  
-        /// Success: return an S3Response object with status 204. 
+        /// Success: send an S3Response with status 204. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> Delete = null;
+        public Func<S3Context, Task> Delete = null;
 
         /// <summary>
         /// Delete a bucket's tags. 
-        /// Success: return an S3Response object with status 204. 
+        /// Success: send an S3Response with status 204. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> DeleteTags = null;
+        public Func<S3Context, Task> DeleteTags = null;
 
         /// <summary>
         /// Delete a bucket's website configuration. 
-        /// Success: return an S3Response object with status 204. 
+        /// Success: send an S3Response with status 204. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> DeleteWebsite = null;
+        public Func<S3Context, Task> DeleteWebsite = null;
 
         /// <summary>
         /// Check for the existence of a bucket.
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Exists = null;
+        public Func<S3Context, Task> Exists = null;
 
         /// <summary>
         /// Enumerate a bucket.  
-        /// Success: return an S3Response object with status 200 and a ListBucketResult object as its data.
+        /// Success: send an S3Response with status 200 and a ListBucketResult object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Read = null;
+        public Func<S3Context, Task> Read = null;
 
         /// <summary>
         /// Read a bucket's access control policy.
-        /// Success: return an S3Response object with status 200 and an AccessControlPolicy object as its data.
+        /// Success: send an S3Response with status 200 and an AccessControlPolicy object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadAcl = null;
+        public Func<S3Context, Task> ReadAcl = null;
 
         /// <summary>
         /// Retrieve logging configuration for this bucket.
-        /// Success: return an S3Response object with status 200 and BucketLoggingStatus in the body.
+        /// Success: send an S3Response with status 200 and BucketLoggingStatus in the body.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadLogging = null;
+        public Func<S3Context, Task> ReadLogging = null;
 
         /// <summary>
         /// Retrieve location (region) constraint from the server for this bucket.
-        /// Success: return an S3Response object with status 200 and LocationConstraint in the body.
+        /// Success: send an S3Response with status 200 and LocationConstraint in the body.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadLocation = null;
+        public Func<S3Context, Task> ReadLocation = null;
 
         /// <summary>
         /// Read a bucket's tags. 
-        /// Success: return an S3Response object with status 200 and a Tagging object as its data. 
+        /// Success: send an S3Response with status 200 and a Tagging object as its data. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadTags = null;
+        public Func<S3Context, Task> ReadTags = null;
 
         /// <summary>
         /// Get a list of object versions in the bucket. 
-        /// Success: return an S3Response object with status 200 and a ListVersionsResult object as its data.
+        /// Success: send an S3Response with status 200 and a ListVersionsResult object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadVersions = null;
+        public Func<S3Context, Task> ReadVersions = null;
 
         /// <summary>
         /// Get a bucket's versioning policy.  
-        /// Success: return an S3Response object with status 200 and a VersioningConfiguration object as its data.
+        /// Success: send an S3Response with status 200 and a VersioningConfiguration object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadVersioning = null;
+        public Func<S3Context, Task> ReadVersioning = null;
 
         /// <summary>
         /// Get a bucket's website configuration.  
-        /// Success: return an S3Response object with status 200 and a WebsiteConfiguration object as its data.
+        /// Success: send an S3Response with status 200 and a WebsiteConfiguration object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadWebsite = null;
+        public Func<S3Context, Task> ReadWebsite = null;
 
         /// <summary>
         /// Write a bucket.  
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Write = null;
+        public Func<S3Context, Task> Write = null;
 
         /// <summary>
         /// Write an ACL to a bucket, deleting the previous ACL.
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteAcl = null;
+        public Func<S3Context, Task> WriteAcl = null;
 
         /// <summary>
         /// Write logging configuration to a bucket, deleting the previous configuration.
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteLogging = null;
+        public Func<S3Context, Task> WriteLogging = null;
 
         /// <summary>
         /// Write tags to a bucket, deleting the previous tags.
-        /// Success: return an S3Response object with status 204.
+        /// Success: send an S3Response with status 204.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteTags = null;
+        public Func<S3Context, Task> WriteTags = null;
 
         /// <summary>
         /// Set a bucket's versioning policy.  
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteVersioning = null;
+        public Func<S3Context, Task> WriteVersioning = null;
 
         /// <summary>
         /// Set a bucket's website configuration.  
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteWebsite = null;
+        public Func<S3Context, Task> WriteWebsite = null;
 
         #endregion