+++ S3ServerInterface/Callbacks/ObjectCallbacks.cs
@@ -15,93 +15,93 @@ namespace S3ServerInterface.Callbacks
 
         /// <summary>
         /// Check for the existence of an object.  
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response object with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Exists = null;
+        public Func<S3Context, Task> Exists = null;
 
         /// <summary>
         /// Write an object.
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response object with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Write = null;
+        public Func<S3Context, Task> Write = null;
 
         /// <summary>
         /// Write an object's access control list, replacing the previous ACL.  
-        /// Success: return an S3Response object with status 200. 
+        /// Success: send an S3Response object with status 200. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteAcl = null;
+        public Func<S3Context, Task> WriteAcl = null;
 
         /// <summary>
         /// Write tags to an object, replacing the previous tags.
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response object with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteTags = null;
+        public Func<S3Context, Task> WriteTags = null;
 
         /// <summary>
         /// Write a legal hold status to an object. 
-        /// Success: return an S3Response object with status 200.
+        /// Success: send an S3Response object with status 200.
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteLegalHold = null;
+        public Func<S3Context, Task> WriteLegalHold = null;
 
         /// <summary>
         /// Write a retention status to an object.
-        /// Success: return an S3Response object with status 200. 
+        /// Success: send an S3Response object with status 200. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> WriteRetention = null;
+        public Func<S3Context, Task> WriteRetention = null;
 
         /// <summary>
         /// Read an object.  
-        /// Success: return an S3Response object with status 200 and the object data.
+        /// Success: send an S3Response object with status 200 and the object data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Read = null;
+        public Func<S3Context, Task> Read = null;
 
         /// <summary>
         /// Read an object's access control list.  
-        /// Success: return an S3Response object with status 200 and an AccessControlPolicy as its data.
+        /// Success: send an S3Response object with status 200 and an AccessControlPolicy as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadAcl = null;
+        public Func<S3Context, Task> ReadAcl = null;
 
         /// <summary>
         /// Read a range of bytes from an object.  
-        /// Success: return an S3Response object with status 200 and the range of bytes as its data. 
+        /// Success: send an S3Response object with status 200 and the range of bytes as its data. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadRange = null;
+        public Func<S3Context, Task> ReadRange = null;
 
         /// <summary>
         /// Read an object's tags.  
-        /// Success: return an S3Response object with status 200 and a Tagging object as its data.
+        /// Success: send an S3Response object with status 200 and a Tagging object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadTags = null;
+        public Func<S3Context, Task> ReadTags = null;
 
         /// <summary>
         /// Read an object's legal hold status.
-        /// Success: return an S3Response object with status 200 and a LegalHold object as its data. 
+        /// Success: send an S3Response object with status 200 and a LegalHold object as its data. 
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadLegalHold = null;
+        public Func<S3Context, Task> ReadLegalHold = null;
 
         /// <summary>
         /// Read an object's retention status.  
-        /// Success: return an S3Response object with status 200 and a Retention object as its data.
+        /// Success: send an S3Response object with status 200 and a Retention object as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> ReadRetention = null;
+        public Func<S3Context, Task> ReadRetention = null;
 
         /// <summary>
         /// Delete an object. 
-        /// Success: return an S3Response object with status 204.
+        /// Success: send an S3Response object with status 204.
         /// </summary>
-        public Func<S3Request, S3Response, Task> Delete = null;
+        public Func<S3Context, Task> Delete = null;
 
         /// <summary>
         /// Delete an object's tags.  
-        /// Success: return an S3Response object with status 204.
+        /// Success: send an S3Response object with status 204.
         /// </summary>
-        public Func<S3Request, S3Response, Task> DeleteTags = null;
+        public Func<S3Context, Task> DeleteTags = null;
 
         /// <summary>
         /// Delete multiple objects.  
-        /// Success: return an S3Response object with status 200 and a DeleteResult as its data.
+        /// Success: send an S3Response object with status 200 and a DeleteResult as its data.
         /// </summary>
-        public Func<S3Request, S3Response, Task> DeleteMultiple = null;
+        public Func<S3Context, Task> DeleteMultiple = null;
 
         #endregion